Big commit: Use lowerer and bytecode
Problem: The scoping system is hacky and should be improved. Functions
do not have access to any variable outside its own scope. This raises
the issue that functions can not call functions unless it is passed
as an argument or declared within the function. This is also means that
recursion is broken