`BadNode` now stores a `TextSpan`
During parsing, even if a `BadNode` is created, parsing continues, in
case the bad node is present, and a new parent node requires its
text span to create a new text span it cause underflow due to
subtraction since `BadNode` used to have a default span of (0, 0).

For example:
```rust
 123 + let /*
^- Note this space in the first line */
```
This code will try to produce the following AST:
         +
        / \
       /   \
     123   bad
   (left) (right)

While producing the span for the `+`, it calls
`TextSpan::from_spans(left, right)` and since the bad has start and
length as 0, it has end = 0. So while calculating length of the binary
node, 0 - left.start occurs (in this case 0 - 1), which causes a panic
in debug mode.

The space was only to make the start > 0, since 0 - 0 will not panic