Change SourceText to be generic over types that implement TextBase
This change was done to allow `&[String]` to be lexed directly since
lexing each line separately can cause erroneous tokens. For example:
```
let a = "Hello
{World!"
```

While executing, this would be valid code and run properly, including
the new line in the string. However, while getting indent, the string
"Hello\n{World!" would not be taken as one string, and the `{` would be
counted as an OpenBrace token which leads to the wrong indent. Also it
would be linted incorrectly, since the quotes occur in different lines.
This commit doesn't fix this **yet** since the language interface only
provides the current line. It will be fixed in the next commit.