Proper representation of inbuilt functions.
Functions are now essentially of two kinds:
- `AnilangFn`:
  Normal functions which are defined within anilang itself.
- `NativeFn`:
  Functions which are basically bindings to those written in rust.

Currently, native functions can only be called using `CALL_INBUILT`, but
that will be remove in the next commit, and instead there will be a std
scope that contains these native functions.

NOTE: Native functions cannot be serialized or deserialized, so they
must not occur in the bytecode directly, and must only be provided by
the runtime.