Added try_unwrap
Gc now has an associated method called try_unwrap which will return the
inner T if it is the last `Gc`. This has added a fat pointer + a usize
more overhead within the inner, since an actual reference count and a
pointer to the previous in the linked list is required.

Making it a doubly linked list allows try_unwrap to easily remove the
GcInner without having to traverse the list.